[EuiModal] Moving close icon to modal_header and pass onClose as a prop through cloneElement
